[503] adding all template tests for residnets table chart
